core: drop ScalarQuery.viaLegacy; resolve pendingJoins in builders
ScalarQuery and ColumnQuery now resolve pendingJoins (joins / whereMissing)
to flat \$in / \$notIn filters before lowering, so the connector path
sees a clean spec without unresolved joins. SQL connectors no longer
need to fall back to queryWithJoins on pendingJoins+aggregate combos —
they throw if pendingJoins arrive at queryScoped. Drop viaLegacy and
extract the shared scope-resolution helpers (builderScopeBase,
resolvePendingJoinsToScope, resolveParentScopesToFilter) into
query/scope.ts so CollectionQuery / ScalarQuery / ColumnQuery share them.
