+++ packages/core/src/query/ScalarQuery.ts
@@ -2,31 +2,9 @@ import type { Dict, Filter, KeyType, Projection } from '../types.js';
 import type { QueryState } from './QueryState.js';
 import { lower, resolveSubqueryFilters } from './lower.js';
 import { PersistenceError } from '../errors.js';
-import { NullConnector } from '../Model.js';
+import { resolvePendingJoinsToScope } from './scope.js';
 
 type ModelLike = { tableName: string; keys: Dict<KeyType> };
-
-/**
- * Mirrors CollectionQuery's makeScopedSubclass — projects QueryState onto a
- * temp Model subclass so the legacy aggregate path (`Model.count()` /
- * `.sum()` / etc., which read `this.<x>` static properties) operates on
- * the chained scope.
- */
-function makeStateProjectedSubclass(M: any, state: QueryState): any {
-  return class extends (M as any) {
-    static filter = state.filter;
-    static order = [...state.order];
-    static limit = state.limit;
-    static skip = state.skip;
-    static selectedFields = state.selectedFields;
-    static selectedIncludes = [...state.selectedIncludes];
-    static includeStrategy = state.includeStrategy;
-    static pendingJoins = [...state.pendingJoins];
-    static havingPredicate = state.havingPredicate;
-    static softDelete = state.softDelete;
-    static connector = state.nullScoped ? new NullConnector() : (M as any).connector;
-  };
-}
 
 export class ScalarQuery<T = unknown> implements PromiseLike<T> {
   protected memo: Promise<T> | undefined;
@@ -65,18 +43,26 @@ export class ScalarQuery<T = unknown> implements PromiseLike<T> {
     if (!this.memo) {
       this.memo = (async () => {
         if (this.state.nullScoped) return this.emptyResult();
-        // pendingJoins (from `joins(...)` / `whereMissing(...)`) are resolved
-        // by Model.modelScope() into concrete `$in` / `$notIn` filters; defer
-        // to the legacy aggregate path so we get that resolution for free.
+        // Resolve `pendingJoins` (`joins(...)` / `whereMissing(...)`) to
+        // concrete `$in` / `$notIn` filters before lowering, so the connector
+        // sees a flat scope — aggregates / column projections never go through
+        // queryWithJoins, which would discard the projection.
+        let stateForLower = this.state;
         if (this.state.pendingJoins.length > 0) {
-          return (await this.viaLegacy()) as T;
+          const resolvedScope = await resolvePendingJoinsToScope(this.model as any, this.state);
+          stateForLower = {
+            ...this.state,
+            filter: resolvedScope.filter,
+            pendingJoins: [],
+          };
         }
-        const resolvedFilter = await resolveSubqueryFilters(this.state.filter);
+        const resolvedFilter = await resolveSubqueryFilters(stateForLower.filter);
         const scopedFilter = this.applyImplicitScopes(resolvedFilter);
-        const builderForLower =
-          scopedFilter !== this.state.filter
-            ? new (this.constructor as any)(this.model, { ...this.state, filter: scopedFilter }, this.projection)
-            : this;
+        const builderForLower = new (this.constructor as any)(
+          this.model,
+          { ...stateForLower, filter: scopedFilter },
+          this.projection,
+        );
         const spec = lower(builderForLower, this.projection);
         const M = this.model as any;
         const connector = M.connector;
@@ -94,40 +80,6 @@ export class ScalarQuery<T = unknown> implements PromiseLike<T> {
       })();
     }
     return this.memo;
-  }
-
-  /**
-   * Fall back to Model.modelScope() + the connector's native aggregate /
-   * select. Used when the chain has pendingJoins (`joins(...)` /
-   * `whereMissing(...)`) — modelScope() resolves those into `$in` / `$notIn`
-   * filters, which baseQueryScoped doesn't know how to do on its own.
-   * We bypass `Sub.count()` / `Sub.sum()` here because those now forward
-   * back into ScalarQuery, which would re-enter this fallback infinitely.
-   */
-  private async viaLegacy(): Promise<unknown> {
-    const M = this.model as any;
-    const Sub = makeStateProjectedSubclass(M, this.state);
-    const scope = await Sub.modelScope();
-    const connector = Sub.connector;
-    if (typeof this.projection === 'object' && this.projection.kind === 'aggregate') {
-      const { op, column } = this.projection;
-      if (op === 'count') return connector.count(scope);
-      if (column === undefined) {
-        throw new PersistenceError(
-          `Aggregate '${op}' requires a column; received undefined.`,
-        );
-      }
-      const value = await connector.aggregate(scope, op as any, column);
-      if (op === 'sum') return value ?? 0;
-      return value;
-    }
-    if (typeof this.projection === 'object' && this.projection.kind === 'column') {
-      // Single-column scalar selection (called from InstanceQuery.pluck) — take
-      // the first row's column value via the connector's select.
-      const items = await connector.select(scope, this.projection.column);
-      return (items[0] as Dict<any> | undefined)?.[this.projection.column];
-    }
-    throw new PersistenceError(`Unsupported projection in legacy fallback: ${JSON.stringify(this.projection)}`);
   }
 
   // Empty-set semantics for the nullScoped short-circuit:
