+++ packages/knex-connector/src/KnexConnector.ts
@@ -232,25 +232,14 @@ export class KnexConnector implements Connector {
    * subqueries — one per `parentScope`. The leaf builder runs against the
    * target table; each parent scope projects its `link.parentColumn` and
    * gates the next level (or the leaf) via that scope's `link.childColumn`.
-   * Falls back to `queryWithJoins` when `pendingJoins` are present.
+   * Builders resolve `pendingJoins` to `$in` / `$notIn` filters before calling
+   * this method, so the connector only sees a flat scope.
    */
   async queryScoped(spec: QueryScopedSpec): Promise<unknown> {
-    // pendingJoins → defer to queryWithJoins (already merges parent-scope IN
-    // filters into the parent filter at the Model layer when present).
-    // TODO: queryWithJoins always returns rows; if a non-'rows' projection
-    // is requested alongside pendingJoins, projection is dropped. Combining
-    // joins with count/sum/pluck isn't yet exercised; revisit when needed.
     if (spec.pendingJoins.length > 0) {
-      return this.queryWithJoins({
-        parent: {
-          tableName: spec.target.tableName,
-          filter: spec.filter,
-          order: spec.order,
-          limit: spec.limit,
-          skip: spec.skip,
-        },
-        joins: spec.pendingJoins,
-      });
+      throw new PersistenceError(
+        `KnexConnector.queryScoped expects pendingJoins to be resolved upstream; received ${spec.pendingJoins.length} unresolved join(s). Use the CollectionQuery / ScalarQuery / ColumnQuery builders to materialise scopes with joins.`,
+      );
     }
 
     let builder = this.table(spec.target.tableName);
